Extract executeAllChildren method and related calls into VintageExecutor
Issue: #4238
